restrict for admin

+++ src/View/composant/detail.php
@@ -6,10 +6,12 @@
         echo "<p> Prix = " . htmlspecialchars($composant->getPrix()) . " </p>\n";
         echo "<p> ImgPath = " . htmlspecialchars($composant->getImgPath()) . " </p>\n";
         echo "<p> Description = " . htmlspecialchars($composant->getDescription()) . " </p>\n";
-        echo "<div class='mt-5'>";
+        if ($estAdmin) {
+            echo "<div class='mt-5'>";
             echo "<p><a href='frontController.php?controller=composant&action=delete&id=" . rawurlencode($composant->getId()) . "' > supprimer </a></p>";
             echo "<p><a href='frontController.php?controller=composant&action=update&id=" . rawurlencode($composant->getId()) . "' > modifier </a></p>";
-        echo "</div>";
+            echo "</div>";
+        }
     ?>
 
 </article>
